+++ src/base/material.inject.js
@@ -1,0 +1,6 @@
+wrap(() => {
+  const el = document.createElement('link');
+  el.href = 'https://fonts.googleapis.com/icon?family=Material+Icons';
+  el.rel = 'stylesheet';
+  document.head.append(el);
+});