+++ src/base/styles/general.js
@@ -1,0 +1,3 @@
+style.add(
+  '.clickable { cursor: pointer; }',
+);